requirements.txt
requests
moviepy
openai
pillow
pexels-api
python-dotenv
fastapi
uvicorn
